requirements.txt
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.2
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-bootstrap-datepicker==1.4.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
html5lib==1.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
xlwt==1.3.0
